Updates for build process
This updates the go.sum to more current versions

This alters the create_release.sh to fix a fault in the tarball being
created before the path definitely existed.

It alters an assumption about the users environment in create_release.sh
by forcing the GO111MODULE to on as the auto value might not result in a
module environment which is incompatible with dependencies.

It also reverts changes I made to my branch in the Makefile to revert it
to match master.